refactor components and utils into lib folder
