+++ src/lib/bookings/calendar/components/header/ViewSwitcher.svelte
