+++ src/lib/bookings/components/InfoItem.svelte
